require trailing slash on urls
By default, when Django reaches the end of the urlconf and still hasn't
made a match, it will redirect to the same url with a '/' appended,
if there wasn't one there. By not requiring a trailing slash, we break
this behaviour and interfere with the non-trailing-slashed versions
of URLs higher up the urlconf.